refactor(home): extract state and event
